Renaming as using abstract class not interface
