+++ src/Config/ObjectModel/BaseRunTimeConfigLoader.cs
@@ -6,16 +6,15 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using Azure.DataApiBuilder.Config.Converters;
 using Azure.DataApiBuilder.Config.NamingPolicies;
-using Azure.DataApiBuilder.Config.ObjectModel;
 using Microsoft.Extensions.Logging;
 
-namespace Azure.DataApiBuilder.Config.Interfaces
+namespace Azure.DataApiBuilder.Config.ObjectModel
 {
-    public abstract class IRunTimeConfigLoader
+    public abstract class BaseRunTimeConfigLoader
     {
         protected readonly string? _connectionString;
 
-        public IRunTimeConfigLoader(string? connectionString = null)
+        public BaseRunTimeConfigLoader(string? connectionString = null)
         {
             _connectionString = connectionString;
         }
